Speed up incremental metrics evaluation by >60x and implement a global AP metric.
I removed the top_k approximation which is no longer needed.
I also simplified/improved some code or documentation as I came across it.

The final metrics computation step which samples the PR curve is still
pretty slow. The total computation time is now 30 minutes, with most of it
being in the final step.

Change-Id: I03cbaeb1e31f46a8f2b39866e899c696ef4711ee